Convert components to single file components
Controllers converted:
* projects/annotate
* projects/edit
* scenes/scenes
* drawToolbar
* annotateSidebar
