--- app-frontend/src/app/components/map/annotateToolbar/annotateToolbar.component.js
@@ -1,16 +1,0 @@
-import annotateToolbarTpl from './annotateToolbar.html';
-
-const annotateToolbar = {
-    templateUrl: annotateToolbarTpl,
-    controller: 'AnnotateToolbarController',
-    bindings: {
-        mapId: '@',
-        disableToolbarAction: '<',
-        bulkMode: '<',
-        onDrawingCanceled: '&',
-        onShapeCreating: '&',
-        onShapeCreated: '&'
-    }
-};
-
-export default annotateToolbar;
