--- app-frontend/src/app/components/projects/annotateSidebarItem/annotateSidebarItem.component.js
@@ -1,21 +1,0 @@
-import annotateSidebarItemTpl from './annotateSidebarItem.html';
-
-const annotateSidebarItem = {
-    templateUrl: annotateSidebarItemTpl,
-    controller: 'AnnotateSidebarItemController',
-    bindings: {
-        annotation: '<',
-        labelInputs: '<',
-        editId: '<',
-        disableSidebarAction: '<',
-        onCloneAnnotation: '&',
-        onUpdateAnnotationStart: '&',
-        onDeleteAnnotation: '&',
-        onUpdateAnnotationFinish: '&',
-        onCancelUpdateAnnotation: '&',
-        onBulkCreate: '&',
-        onQaChecked: '&'
-    }
-};
-
-export default annotateSidebarItem;
